requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
beautifulsoup4
streamlit
requests
faiss-cpu
google-generativeai==0.8.1
html5lib
